DevOps: add docker-compose file for dev environment
With this setup the app is running in an isolated environment, with all
dependencies encapsulated in Docker containers. The Compose file
included here is configured for the development environment, another
Compose file should be used for production with different environment
variables (e.g. database password) and a strategy for data backup.