Fix QuestionFilePondField to reset fileOrigin
After the user selects a file for a question, fileOrigin becomes 'input'.
Even after saving, and going to the details page, fileOrigin continues to
be 'input', because it's declared in the module scope. So we need to
reset it to 'local' when the file is removed from the pond after the form
is submitted. Otherwise, if he/she goes to the edit page again, the image
he/she just uploaded doesn't show up, because fileOrigin is 'input'.