Show custom Snackbar after deleting a question
This is not working because the Snackbar's “open” state is lost after redirecting
to the list page.
The Snackbar is mounted in the SafeDeleteIconButton component, which is
unmounted when redirecting.
One possible solution is to keep the Snackbar's state in a custom Next.js' App
component (`./pages/_app.js`). We could use React's Context API to pass it down,
together with a function to update it.