Set up TypeScript import aliases (#140)
* Change start script to use lerna run --parallel

It's seems better than npm-run-all.

* Add comments about tsconfig options

* Add launch.json to .vscode

"Next: Node" and "Jest Tests" included.

* Move api & web to packages directory

* Fix directory structure in config files

* Fix directory structure in CircleCI config

* Set up TypeScript project references

https://www.typescriptlang.org/docs/handbook/project-references.html

* Add @olimat/api as dependency of @olimat/web

* Fix build command in CircleCI config

* Disable @typescript-eslint/no-unused-vars rule

* Add '@olimat/api' path alias to api/tsconfig.json

This is for path remapping within the package. To remap
external packages we use Lerna symlinks.

* Ignore '@olimat/api' globally for import/no-unresolved

For some reason eslint can't resolve @olimat/api used in @olimat/web.
It's symlinked by lerna, so it should just work. It's not the best
solution but we're ignoring globally for now.

* Add CI job step for building TS project references

* Set up moduleNameMapper in jest.workspace.js

This is not ideal, but it's what we have for today.

* Add npm script "build:skip-web" for CI

This is getting ugly...

* Add "rimraf dist" to "prebuild" npm script in api

* Fix TypeScript paths after compilation

Why do I have the feeling this import alias thing
is getting out of hand?

* Replace relative imports w/ aliases in API code

* Fix issue with WithDataProps interface

TS was complaining: "[ts] Default export of the module
has or is using private name 'WithDataProps'."

* Add @olimat/web path mapping in web/tsconfig.json

* Replace relative imports w/ aliases in frontend code

* Replace relative imports in frontend tests

... with aliases