Fix unavailable next/router in getDataFromTree()
"Error while running `getDataFromTree` TypeError: Cannot read property
'pathname' of null".

There's a lot of code I added in this commit just to debug what was
going on, but the important part is the usage of `AppTree` instead
of the composed `App` when using `getDataFromTree()`.

The issue is documented here:
https://github.com/vercel/next.js/issues/7731

And solved here:
https://github.com/vercel/next.js/pull/7732